# internal/rest/openapi3.json
{"components":{"requestBodies":{"CreateUpdatePostRequest":{"content":{"application/json":{"schema":{"properties":{"authorId":{"format":"int32","type":"integer"},"text":{"minLength":1,"type":"string"}}}}},"description":"Тело запроса для создания и обновления поста","required":true}},"responses":{"CreateUpdatePostResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"},"result":{"format":"int32","type":"integer"}}}}},"description":"Ответ при успешном создании или обновлении поста"},"ErrorResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"},"result":{"format":"int32","type":"integer"}}}}},"description":"Ответ при ошибке"},"SelectAllPostsResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"},"result":{"items":{"$ref":"#/components/schemas/Post"}}}}}},"description":"Ответ при успешном получении всех постов"},"SelectPostByIdResponse":{"content":{"application/json":{"schema":{"properties":{"error":{"type":"string"},"result":{"$ref":"#/components/schemas/Post"}}}}},"description":"Ответ при успешном получении поста по id"}},"schemas":{"Author":{"properties":{"avatar":{"nullable":true,"type":"string"},"id":{"format":"int32","type":"integer"},"username":{"type":"string"}},"type":"object"},"Comments":{"properties":{"author":{"$ref":"#/components/schemas/Author"},"dates":{"$ref":"#/components/schemas/Dates"},"id":{"format":"int32","type":"integer"},"text":{"type":"string"}},"type":"array"},"Dates":{"properties":{"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"type":"object"},"Post":{"properties":{"author":{"$ref":"#/components/schemas/Author"},"comments":{"$ref":"#/components/schemas/Comments"},"dates":{"$ref":"#/components/schemas/Dates"},"id":{"format":"int32","type":"integer"},"text":{"type":"string"}},"type":"object"}}},"info":{"description":"REST API used for interacting with Art space app","title":"Art space API","version":""},"openapi":"3.0.0","paths":{"/posts":{"get":{"operationId":"SelectAll","responses":{"200":{"$ref":"#/components/responses/SelectAllPostsResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"post":{"operationId":"CreatePost","requestBody":{"$ref":"#/components/requestBodies/CreateUpdatePostRequest"},"responses":{"201":{"$ref":"#/components/responses/CreateUpdatePostResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}},"/posts/{postId}":{"delete":{"operationId":"DeletePost","parameters":[{"in":"path","name":"postId","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/CreateUpdatePostResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"get":{"operationId":"SelectById","parameters":[{"in":"path","name":"postId","required":true,"schema":{"format":"int32","type":"integer"}}],"responses":{"200":{"$ref":"#/components/responses/SelectPostByIdResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}},"put":{"operationId":"UpdatePost","parameters":[{"in":"path","name":"postId","required":true,"schema":{"format":"int32","type":"integer"}}],"requestBody":{"$ref":"#/components/requestBodies/CreateUpdatePostRequest"},"responses":{"200":{"$ref":"#/components/responses/CreateUpdatePostResponse"},"400":{"$ref":"#/components/responses/ErrorResponse"},"500":{"$ref":"#/components/responses/ErrorResponse"}}}}},"servers":[{"description":"Local dev","url":"http://172.25.122.55:8000"}]}